Add files via upload
Basic Preprocessing  is done on the fish dataset  ,this dataset is a record of 7 common different fish species in fish market sales. With this dataset, a predictive model can be performed using machine friendly data and estimate the weight of fish can be predicted. 
So we have  performed basic preprocessing for handling the null values and verifying which imputing strategy is giving better results for predictions.